pyproject.toml
[tool.poetry]
name = "shanyraq"
version = "0.1.0"
description = ""
authors = ["beliyaka98 <the.akniyet.second@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.11.1"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
pyjwt = "^2.8.0"
python-decouple = "^3.8"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
